Create Find and Replace
Find and Replace


Polymens a famous news editor, has daily responsibility of deciding which news stories are printed in the paper. Before that he wants to verify all the words are spelled in a right way. Incase of any misspelled word found, he has to replace it throughout the document without any fail. Polymens needs an application to make his find and replace job easier. 

Implement the given scenario and code accordingly. Refer the samples for read and display the data.

[Search and replace should be case sensitive.  Input can contain words separated by space and full stop.]

Sample Input 1:

Enter the string:

Have a good day

Enter the word to be searched:

good

Enter the word to be replaced:

nice

Sample Output 1:

Have a nice day

Sample Input 2:

Enter the string:

Have a good day

Enter the word to be searched:

nice

Enter the word to be replaced:

nice

Sample Output 2:

The word nice not found



Sample Input 3:

Enter the string:
Google PlayStore is a digital distribution service operated and developed by Google.

Enter the word to be searched:

Google

Enter the word to be replaced:

Yahoo

Sample Output 3:

Yahoo PlayStore is a digital distribution service operated and developed by Yahoo.